+++ codex/views/opds/v2/manifest.py
@@ -15,7 +15,7 @@ from codex.serializers.opds.v2.publication import (
 )
 from codex.views.auth import GroupACLMixin
 from codex.views.opds.const import AUTHOR_ROLES, BLANK_TITLE, MimeType, Rel
-from codex.views.opds.util import get_credits, get_m2m_objects
+from codex.views.opds.metadata import get_credits, get_m2m_objects
 from codex.views.opds.v2.const import HrefData, LinkData
 from codex.views.opds.v2.feed.publications import OPDS2PublicationBaseView